feat/packaging : Correcting packaging for pip compliance
