+++ .github/workflows/latest.yml
@@ -41,7 +41,7 @@ jobs:
               env:
                 SNYK_TOKEN: ${{ secrets.SNYK_TOKEN }}
               with:
-                args: --sarif-file-output=snyk.sarif
+                args: --sarif-file-output=snyk.sarif --file=setup.py
 
             - name: Run pip-audit to check for vulnerabilities
               continue-on-error: true
