+++ .github/workflows/security.yml
@@ -40,7 +40,7 @@ jobs:
               continue-on-error: true
               uses: docker://technogix/terraform-python-awscli:v1.1.0
               with:
-                args: pip-audit requirements-test.txt
+                args: pip-audit -r requirements-test.txt
 
             - name: Upload snyk result to GitHub Code Scanning
               uses: github/codeql-action/upload-sarif@v1
