--- keywords/__init__.py
@@ -1,4 +1,0 @@
-""" List all files in module """
-__all__ = [ "account", "analyzer", "cloudtrail", "cloudwatch", "codecommit", \
-            "config", "directory", "dynamo", "ec2", "ecr", "iam", "kms", "rds", "resourcegroup", \
-            "s3", "sns", "sso", "terraform", "profile", "keepass"]
