+++ technogix_iac_keywords/kms.py
@@ -22,7 +22,7 @@ from robot.api.deco import keyword
 ROBOT = False
 
 # Local includes
-syspath.append(path.normpath(path.join(path.dirname(__file__), '../')))
+syspath.append(path.normpath(path.join(path.dirname(__file__), './')))
 from tools.kms      import KMSTools
 from tools.compare  import compare_dictionaries, remove_type_from_dictionary, remove_type_from_list
